build: remove circular dependencies
