+++ ts/main.ts
@@ -1,7 +1,8 @@
 import { displayCSS, displayHTML } from "./logic/display";
-import { main } from "./logic/main";
+import { generateHexagonSection } from "./logic/main";
 
 // call the function for an initial display of hexagons
-main();
+generateHexagonSection();
+// add event listeners to buttons
 displayCSS();
 displayHTML();
